背景色のデフォルト値を透明の白に変更. バックアアップの不具合を修正.

+++ app/src/main/java/com/pin/recommend/model/entity/translation/AccountExportable.kt
@@ -23,7 +23,7 @@ class AccountExportable{
 
     fun importable(): Account{
         val account = Account()
-        account.id = Account.ACCOUNT_ID.toLong()
+        account.id = Account.ACCOUNT_ID
         toolbarTextColor?.let {
             try{
                 account.toolbarTextColor = Color.parseColor(it)
